proper handling of recursion for rename-annotations and remove-annotations

+++ ChildProject/annotations.py
@@ -377,8 +377,31 @@ class AnnotationManager:
         self.annotations = pd.concat([self.annotations, imported], sort = False)
         self.annotations.to_csv(os.path.join(self.project.path, 'metadata/annotations.csv'), index = False)
 
-    def remove_set(self, annotation_set):
+    def get_subsets(self, annotation_set, recursive = False):
+        subsets = []
+
+        path = os.path.join(self.project.path, 'annotations', annotation_set)
+        candidates = list(set(os.listdir(path)) - {'raw', 'converted'})
+        for candidate in candidates:
+            if os.path.exists(os.path.join(path, candidate, 'raw')):
+                subset = os.path.join(annotation_set, candidate)
+                subsets.append(subset)
+                if recursive:
+                    subsets.extend(self.get_subsets(subset))
+
+        return subsets
+            
+
+    def remove_set(self, annotation_set, recursive = False):
         self.read()
+
+        subsets = []
+        if recursive:
+            subsets = self.get_subsets(annotation_set, recursive = False)
+
+        for subset in subsets:
+            self.remove_set(subset, recursive = recursive)
+
         path = os.path.join(self.project.path, 'annotations', annotation_set, 'converted')
 
         try:
@@ -409,20 +432,13 @@ class AnnotationManager:
             raise Exception("set '{}' have no indexed annotation, aborting. use --ignore_errors to force")
 
         subsets = []
-        
         if recursive:
-            candidates = list(set(os.listdir(current_path)) - {'raw', 'converted'})
-            for candidate in candidates:
-                if os.path.exists(os.path.join(current_path, candidate, 'raw')):
-                    subsets.append({
-                        'current': os.path.join(annotation_set, candidate),
-                        'new': os.path.join(new_set, candidate)
-                    })
+            subsets = self.get_subsets(annotation_set, recursive = False)
 
         for subset in subsets:
             self.rename_set(
-                annotation_set = subset['current'],
-                new_set = subset['new'],
+                annotation_set = subset,
+                new_set = re.sub(r"^{}/".format(re.escape(annotation_set)), os.path.join(new_set, ''), subset),
                 recursive = recursive,
                 ignore_errors = ignore_errors
             )
@@ -430,11 +446,14 @@ class AnnotationManager:
         os.makedirs(new_path, exist_ok = True)
 
         shutil.move(os.path.join(current_path, 'raw'), os.path.join(new_path, 'raw'))
-        shutil.move(os.path.join(current_path, 'converted'), os.path.join(new_path, 'converted'))
+
+        if os.path.exists(os.path.join(current_path, 'converted')):
+            shutil.move(os.path.join(current_path, 'converted'), os.path.join(new_path, 'converted'))
 
         self.annotations.loc[(self.annotations['set'] == annotation_set), 'annotation_filename'] = self.annotations.loc[(self.annotations['set'] == annotation_set), 'annotation_filename']\
             .str.replace(r"^{}/".format(re.escape(annotation_set)), repl = os.path.join(new_set, ''), regex = True)
-        self.annotations.loc[(self.annotations['set'] == annotation_set), 'set'] = new_set  
+        self.annotations.loc[(self.annotations['set'] == annotation_set), 'set'] = new_set
+
         self.annotations.to_csv(os.path.join(self.project.path, 'metadata/annotations.csv'), index = False)
